float

+++ src/planning/obstacle_distance_grid.cpp
@@ -100,7 +100,7 @@ void ObstacleDistanceGrid::enqueue_obstacle_cells(ObstacleDistanceGrid& grid, st
 void ObstacleDistanceGrid::expand_node(const DistanceNode& node, ObstacleDistanceGrid& grid, std::priority_queue<DistanceNode>& searchQueue) {
     const int xDeltas[8] = { 1, 1, 1, 0, 0, -1, -1, -1 };
     const int yDeltas[8] = { 0, -1, -1, -1, 1, 1, -1, 0 };
-    const int distances[8] = { 1, std::sqrt(2), 1, std::sqrt(2), 1, std::sqrt(2), 1, std::sqrt(2) };
+    const float distances[8] = { 1, std::sqrt(2), 1, std::sqrt(2), 1, std::sqrt(2), 1, std::sqrt(2) };
 
     for (int i = 0; i < 8; i++) {
         cell_t adjacentCell(node.cell.x + xDeltas[i], node.cell.y + yDeltas[i]);
